refactor: NewsSection에서 뉴스 항목 렌더링 부분을 NewsItem으로 분리 + 에러 처리 추가
